pyproject.toml
[project]
name = "pylearn"
version = "0.1.0"
description = ""
authors = [
    {name = "JustKillEnemies",email = "hewlett74@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "tortoise-orm (>=0.25.1,<0.26.0)",
    "loguru (==0.7.2)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "fastapi (>=0.115.14,<0.116.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "aerich (>=0.9.1,<0.10.0)",
    "tomlkit (>=0.13.3,<0.14.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "annotated-types (>=0.7.0,<0.8.0)",
    "fastapi-admin (>=1.0.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyopenssl (>=25.1.0,<26.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)"
]


[tool.poetry]
packages = [{include = "src"}]


[tool.poetry.scripts]
start = "src.cli:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "src.core.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
